requirements_simulator.txt
cycler>=0.10.0
kiwisolver>= 1.2.0
matplotlib>=3.2.1
numpy>=1.18.2
opencv-python-headless==4.4.0.46
pip>=19.0.3
pyparsing>=2.4.7
python-dateutil>=2.8.1
scipy>=1.4.1
setuptools>=40.8.0
six>=1.14.0
tqdm>=4.54.1
pytest>=6.2.1
streamlit>=0.74.1
openexr==1.3.2; platform_system == 'Linux'
PySide2>=5.15.2
pydantic>=1.8
pandas>=1.1.5
dataclasses>=0.8; python_version == '3.6'
torch~=1.8
ruamel.yaml~=0.17
semver>=2.13
fasteners >= 0.17.3